application/otel-sdk-django-01/requirements.txt
grpcio-health-checking==1.60.0
grpcio-tools==1.59.2
grpcio==1.60.0
opentelemetry-distro==0.46b0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
python-dotenv==1.0.0
python-json-logger==2.0.7
openfeature-sdk==0.7.0
openfeature-provider-flagd==0.1.5
openfeature-hooks-opentelemetry==0.1.3
psutil==5.9.6 # Importing this will also import opentelemetry-instrumentation-system-metrics when running opentelemetry-bootstrap